+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 
 ###### v0.1.2 - 6/9/17
 * Loggers property reformatted from log map and log list
+
+###### v0.2.0 - 8/16/17
+* Blueprints integration and default exception handling for Marshmallow
+* Do no init_app with flask_restful.Api